+++ src/components/pages/login/Login.jsx
@@ -28,27 +28,6 @@ export const Login = () => {
             <LogoLogin />
             <TitleLogin title="Control Ventas" />
             <FormLogin dataForm={dataForm} />
-            {/* <form className="p-3 mt-3">
-                <article className="form-field d-flex align-items-center">
-                    <span className="far fa-user"></span>
-                    <input
-                        type="text"
-                        name="userName"
-                        id="userName"
-                        placeholder="Username"
-                    />
-                </article>
-                <article className="form-field d-flex align-items-center">
-                    <span className="fas fa-key"></span>
-                    <input
-                        type="password"
-                        name="password"
-                        id="pwd"
-                        placeholder="Password"
-                    />
-                </article>
-                <button className="btn mt-3">Login</button>
-            </form> */}
             <article className="text-center fs-6">
                 <a href="#">Forget password?</a> or <a href="#">Sign up</a>
             </article>
